requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.27.2
Pillow==11.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
pymongo==4.6.1
